ast: Fix ore compiler warnings
Most serious ones caused by d8937a5a929f4 accidentally committing the fully
expanded fmapping.c

+++ libraries/ast/grismmap.h
@@ -37,12 +37,12 @@
 *     License as published by the Free Software Foundation, either
 *     version 3 of the License, or (at your option) any later
 *     version.
-*     
+*
 *     This program is distributed in the hope that it will be useful,
 *     but WITHOUT ANY WARRANTY; without even the implied warranty of
 *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 *     GNU Lesser General Public License for more details.
-*     
+*
 *     You should have received a copy of the GNU Lesser General
 *     License along with this program.  If not, see
 *     <http://www.gnu.org/licenses/>.
@@ -97,7 +97,7 @@ typedef struct AstGrismMap {
    double waver;
    double alpha;
    double g;
-   double m;
+   int m;
    double eps;
    double theta;
    double k1;
